Further Evidence Interview: contract of employment
Add code to populate the ‘further evidence interview’ outcome page (END099) with the need to bring in a contract of employment.

Correct wording of comment to show that following code deals with a contract of employment not ID.

Change routing to use the new ‘bookFurtherEvidenceInterviewMarriageCert’ page (which can handle the additional evidence items) instead of the static ‘bookFurtherEvidenceInterviewMarriageCert’ outcome page.